requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy<2.0
Flask-WTF==1.0.1
argon2-cffi==21.3.0
cryptography==38.0.4
Werkzeug==2.2.2